django_project/requirements.txt
asgiref==3.3.4
Django==3.0.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djongo==1.3.3
Pillow==8.2.0
pymongo==3.11.3
pytz==2021.1
sqlparse==0.2.4
